requirements.txt
asgiref==3.3.1
bleach==3.3.0
Django==3.1.6
django-grappelli==2.14.3
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.13
Faker==6.3.0
importlib-metadata==3.4.0
itsdangerous==1.1.0
jsonfield==3.1.0
Markdown==3.3.3
packaging==20.9
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-memcached==1.59
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
swapper==1.1.2.post1
text-unidecode==1.3
typing-extensions==3.7.4.3
webencodings==0.5.1
zipp==3.4.0
